requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
audioread=2.1.4=py36_2
bzip2=1.0.6=h6d464ef_2
ca-certificates=2017.08.26=h1d4fec5_0
certifi=2017.11.5=py36hf29ccca_0
cffi=1.11.2=py36h2825082_0
cudatoolkit=8.0=3
cudnn=6.0.21=cuda8.0_0
cycler=0.10.0=py36h93f1223_0
dbus=1.10.22=h3b5a359_0
decorator=4.1.2=py36hd076ac8_0
expat=2.2.5=he0dffb1_0
ffmpeg=2.8.6=5
fontconfig=2.12.4=h88586e7_1
freetype=2.8=hab7d2ae_1
gettext=0.19.8.1=hd7bead4_3
glib=2.53.6=h5d9569c_2
gst-plugins-base=1.12.2=he3457e5_0
gstreamer=1.12.2=h4f93127_0
icu=58.2=h9c2bf20_1
intel-openmp=2018.0.0=hc7b2577_8
joblib=0.11=py36_0
jpeg=9b=h024ee3a_2
libedit=3.1=heed3624_0
libffi=3.2.1=hd88cf55_4
libgcc=7.2.0=h69d50b8_2
libgcc-ng=7.2.0=h7cc24e2_2
libgfortran-ng=7.2.0=h9f7466a_2
libiconv=1.15=h63c8f33_5
libpng=1.6.32=hbd3595f_4
librosa=0.5.1=py36_1
libstdcxx-ng=7.2.0=h7a57d05_2
libxcb=1.12=hcd93eb1_4
libxml2=2.9.4=h2e8b1d7_6
llvmlite=0.20.0=py36_0
matplotlib=2.1.0=py36hba5de38_0
mkl=2018.0.1=h19d6760_4
nccl=1.3.4=cuda8.0_1
ncurses=6.0=h9df7e31_2
numba=0.35.0=np113py36_10
numpy=1.13.3=py36ha12f23b_0
openssl=1.0.2n=hb7f436b_0
pcre=8.41=hc27e229_1
pip=9.0.1=py36h6c6f9ce_4
pycparser=2.18=py36hf9f622e_1
pyparsing=2.2.0=py36hee85983_1
pyqt=5.6.0=py36h0386399_5
python=3.6.3=h6c0c0dc_5
python-dateutil=2.6.1=py36h88d3b88_1
pytorch=0.2.0=py36cuda8.0cudnn6.0_0
pytz=2017.3=py36h63b9c63_0
qt=5.6.2=h974d657_12
readline=7.0=ha6073c6_4
resampy=0.2.0=py36_1
scikit-learn=0.19.1=py36h7aa7ec6_0
scipy=1.0.0=py36hbf646e7_0
setuptools=36.5.0=py36he42e2e1_0
sip=4.18.1=py36h51ed4ed_2
six=1.11.0=py36h372c433_1
sqlite=3.20.1=hb898158_2
tk=8.6.7=hc745277_3
tornado=4.5.2=py36h1283b2a_0
wheel=0.30.0=py36hfd4bba0_1
x264=20131217=3
xorg-libxau=1.0.8=3
xorg-libxdmcp=1.1.2=3
xz=5.2.3=h55aa19d_2
zlib=1.2.11=ha838bed_2
